refactor(components): [slider] switch to script-setup syntax (#8008)
