+++ packages/components/slider/index.ts
@@ -1,15 +1,8 @@
-import Slider from './src/index.vue'
+import { withInstall } from '@element-plus/utils'
 
-import type { App } from 'vue'
-import type { SFCWithInstall } from '@element-plus/utils'
+import Slider from './src/slider.vue'
 
-Slider.install = (app: App): void => {
-  app.component(Slider.name, Slider)
-}
+export const ElSlider = withInstall(Slider)
+export default ElSlider
 
-const _Slider = Slider as SFCWithInstall<typeof Slider>
-
-export default _Slider
-export const ElSlider = _Slider
-
-export * from './src/slider.type'
+export * from './src/slider'
