+++ packages/components/slider/src/composables/index.ts
@@ -1,0 +1,6 @@
+export * from './use-lifecycle'
+export * from './use-marks'
+export * from './use-slide'
+export * from './use-slider-button'
+export * from './use-stops'
+export * from './use-watch'
