+++ packages/components/slider/src/composables/use-slide.ts
@@ -1,32 +1,27 @@
-import { computed, inject, nextTick, ref, shallowRef } from 'vue'
+import { computed, nextTick, ref, shallowRef } from 'vue'
 import {
   CHANGE_EVENT,
   INPUT_EVENT,
   UPDATE_MODEL_EVENT,
 } from '@element-plus/constants'
-import { formContextKey, formItemContextKey } from '@element-plus/tokens'
-import type { CSSProperties, ComponentInternalInstance, Ref } from 'vue'
-import type {
-  ButtonRefs,
-  ISliderButton,
-  ISliderInitData,
-  ISliderProps,
-} from './slider.type'
-import type { FormContext, FormItemContext } from '@element-plus/tokens'
+import { useFormItem } from '@element-plus/hooks'
+import type { CSSProperties, Ref, SetupContext } from 'vue'
+import type { Arrayable } from '@element-plus/utils'
+import type { SliderEmits, SliderInitData, SliderProps } from '../slider'
+import type { ButtonRefs, SliderButtonInstance } from '../button'
 
 export const useSlide = (
-  props: ISliderProps,
-  initData: ISliderInitData,
-  emit: ComponentInternalInstance['emit']
+  props: SliderProps,
+  initData: SliderInitData,
+  emit: SetupContext<SliderEmits>['emit']
 ) => {
-  const elForm = inject(formContextKey, {} as FormContext)
-  const elFormItem = inject(formItemContextKey, {} as FormItemContext)
+  const { form: elForm, formItem: elFormItem } = useFormItem()
 
   const slider = shallowRef<HTMLElement>()
 
-  const firstButton = ref<ISliderButton>()
+  const firstButton = ref<SliderButtonInstance>()
 
-  const secondButton = ref<ISliderButton>()
+  const secondButton = ref<SliderButtonInstance>()
 
   const buttonRefs: ButtonRefs = {
     firstButton,
@@ -34,7 +29,7 @@ export const useSlide = (
   }
 
   const sliderDisabled = computed(() => {
-    return props.disabled || elForm.disabled || false
+    return props.disabled || elForm?.disabled || false
   })
 
   const minValue = computed(() => {
@@ -86,7 +81,7 @@ export const useSlide = (
 
   const getButtonRefByPercent = (
     percent: number
-  ): Ref<ISliderButton | undefined> => {
+  ): Ref<SliderButtonInstance | undefined> => {
     const targetValue = props.min + (percent * (props.max - props.min)) / 100
     if (!props.range) {
       return firstButton
@@ -109,15 +104,17 @@ export const useSlide = (
     return buttonRefs[buttonRefName]
   }
 
-  const setPosition = (percent: number): Ref<ISliderButton | undefined> => {
+  const setPosition = (
+    percent: number
+  ): Ref<SliderButtonInstance | undefined> => {
     const buttonRef = getButtonRefByPercent(percent)
     buttonRef.value!.setPosition(percent)
     return buttonRef
   }
 
-  const setFirstValue = (firstValue: number) => {
-    initData.firstValue = firstValue
-    _emit(props.range ? [minValue.value, maxValue.value] : firstValue)
+  const setFirstValue = (firstValue: number | undefined) => {
+    initData.firstValue = firstValue!
+    _emit(props.range ? [minValue.value, maxValue.value] : firstValue!)
   }
 
   const setSecondValue = (secondValue: number) => {
@@ -128,7 +125,7 @@ export const useSlide = (
     }
   }
 
-  const _emit = (val: number | number[]) => {
+  const _emit = (val: Arrayable<number>) => {
     emit(UPDATE_MODEL_EVENT, val)
     emit(INPUT_EVENT, val)
   }
@@ -143,7 +140,7 @@ export const useSlide = (
 
   const handleSliderPointerEvent = (
     event: MouseEvent | TouchEvent
-  ): Ref<ISliderButton | undefined> | undefined => {
+  ): Ref<SliderButtonInstance | undefined> | undefined => {
     if (sliderDisabled.value || initData.dragging) return
     resetSize()
     let newPercent = 0
