+++ packages/tokens/index.ts
@@ -9,6 +9,7 @@ export * from './pagination'
 export * from './radio'
 export * from './row'
 export * from './scrollbar'
+export * from './slider'
 export * from './tabs'
 export * from './upload'
 export * from './experimental-features'
